@@ -1,7 +1,7 @@
 package com.nguyenhoanglam.imagepicker.ui.common;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 
 import com.nguyenhoanglam.imagepicker.ui.imagepicker.ImageLoader;
