@@ -1,7 +1,7 @@
 package com.nguyenhoanglam.sample;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -23,13 +23,11 @@ public class ImageAdapter extends RecyclerView.Adapter<ImageAdapter.ImageViewHol
     private Context context;
     private List<Image> images;
     private LayoutInflater inflater;
-    private RequestOptions options;
 
     public ImageAdapter(Context context) {
         this.context = context;
         inflater = LayoutInflater.from(context);
         images = new ArrayList<>();
-        options = new RequestOptions().placeholder(R.drawable.image_placeholder).error(R.drawable.image_placeholder);
     }
 
     @Override
@@ -40,11 +38,11 @@ public class ImageAdapter extends RecyclerView.Adapter<ImageAdapter.ImageViewHol
     @Override
     public void onBindViewHolder(ImageViewHolder holder, int position) {
         final Image image = images.get(position);
-
         Glide.with(context)
                 .load(image.getPath())
-                .apply(options)
+                .apply(new RequestOptions().placeholder(R.drawable.image_placeholder).error(R.drawable.image_placeholder))
                 .into(holder.imageView);
+
     }
 
     @Override
