@@ -53,7 +53,9 @@ public class Config implements Parcelable {
     private String imageTitle;
     private String limitMessage;
     private SavePath savePath;
+    private boolean isAlwaysShowDoneButton;
     private boolean isKeepScreenOn;
+    private int requestCode;
     private ArrayList<Image> selectedImages;
 
 
@@ -78,7 +80,9 @@ public class Config implements Parcelable {
         this.imageTitle = in.readString();
         this.limitMessage = in.readString();
         this.savePath = in.readParcelable(SavePath.class.getClassLoader());
+        this.isAlwaysShowDoneButton = in.readByte() != 0;
         this.isKeepScreenOn = in.readByte() != 0;
+        this.requestCode = in.readInt();
         this.selectedImages = in.createTypedArrayList(Image.CREATOR);
     }
 
@@ -139,7 +143,7 @@ public class Config implements Parcelable {
 
     public int getBackgroundColor() {
         if (TextUtils.isEmpty(backgroundColor)) {
-            return Color.parseColor("#FFFFFF");
+            return Color.parseColor("#212121");
         }
         return Color.parseColor(backgroundColor);
     }
@@ -236,12 +240,28 @@ public class Config implements Parcelable {
         this.savePath = savePath;
     }
 
+    public boolean isAlwaysShowDoneButton() {
+        return isAlwaysShowDoneButton;
+    }
+
+    public void setAlwaysShowDoneButton(boolean isAlwaysShowDoneButton) {
+        this.isAlwaysShowDoneButton = isAlwaysShowDoneButton;
+    }
+
     public boolean isKeepScreenOn() {
         return isKeepScreenOn;
     }
 
     public void setKeepScreenOn(boolean keepScreenOn) {
         isKeepScreenOn = keepScreenOn;
+    }
+
+    public int getRequestCode() {
+        return requestCode;
+    }
+
+    public void setRequestCode(int requestCode) {
+        this.requestCode = requestCode;
     }
 
     public ArrayList<Image> getSelectedImages() {
@@ -276,7 +296,9 @@ public class Config implements Parcelable {
         dest.writeString(this.imageTitle);
         dest.writeString(this.limitMessage);
         dest.writeParcelable(this.savePath, flags);
+        dest.writeByte(this.isAlwaysShowDoneButton ? (byte) 1 : (byte) 0);
         dest.writeByte(this.isKeepScreenOn ? (byte) 1 : (byte) 0);
+        dest.writeInt(this.requestCode);
         dest.writeTypedList(this.selectedImages);
     }
 }
